Add script to prune dictionary based on past NYT answers.
This script parses the raw answers from the nytbee.com scrape and tries to identify valid puzzles. For every puzzle it identifies, it runs our own solver and compares it against the nytbee.com list to find words to add/delete.

We could just scrape nytbee.com again but I would rather not. This is good enough.