Refactor prefix tree to be structured like the radix tree.
This reduces the time taken to almost the same. This is because a nested dictionary structure means we need to run less "dictionary access" operations. The only reason prefix tree takes more time than radix tree now is because there is an extra slice operation in prefix-tree to check the letter is valid. So radix tree just happens to be better suited to this particular task.